move previous test, add failing tests to populate
